Removed getEvents from chartUtils
