+++ src/components/charts/dailyTrendChart/dailyTrendChart.tsx
@@ -9,18 +9,21 @@ import {
   ChartLegendTooltip,
   ChartLine,
   createContainer,
+  getInteractiveLegendEvents,
 } from '@patternfly/react-charts';
 import { Title } from '@patternfly/react-core';
 import { default as ChartTheme } from 'components/charts/chartTheme';
 import { getCostRangeString, getDateRange } from 'components/charts/common/chartDatumUtils';
 import {
   ChartSeries,
+  getChartNames,
   getDomain,
-  getEvents,
   getLegendData,
   getTooltipLabel,
   initHiddenSeries,
   isDataAvailable,
+  isDataHidden,
+  isSeriesHidden,
 } from 'components/charts/common/chartUtils';
 import getDate from 'date-fns/get_date';
 import i18next from 'i18next';
@@ -302,6 +305,20 @@ class DailyTrendChart extends React.Component<DailyTrendChartProps, State> {
     return currentDate > 0 || previousDate > 0 ? Math.max(currentDate, forecastDate, previousDate) : 31;
   }
 
+  // Returns onMouseOver, onMouseOut, and onClick events for the interactive legend
+  private getEvents() {
+    const { hiddenSeries, series } = this.state;
+
+    const result = getInteractiveLegendEvents({
+      chartNames: getChartNames(series),
+      isDataHidden: data => isDataHidden(series, hiddenSeries, data),
+      isHidden: index => isSeriesHidden(hiddenSeries, index),
+      legendName: 'legend',
+      onLegendClick: props => this.handleLegendClick(props.index),
+    } as any); // Todo: remove "as any" when PatternFly's isDataHidden becomes available
+    return result;
+  }
+
   private getLegend = () => {
     const { legendItemsPerRow } = this.props;
     const { hiddenSeries, series, width } = this.state;
@@ -388,7 +405,7 @@ class DailyTrendChart extends React.Component<DailyTrendChartProps, State> {
             <Chart
               containerComponent={container}
               domain={domain}
-              events={getEvents(series, hiddenSeries, this.handleLegendClick, true)}
+              events={this.getEvents()}
               height={height}
               legendAllowWrap
               legendComponent={this.getLegend()}
