+++ src/components/charts/historicalCostChart/historicalCostChart.tsx
@@ -7,6 +7,7 @@ import {
   ChartLegend,
   ChartLegendTooltip,
   createContainer,
+  getInteractiveLegendEvents,
 } from '@patternfly/react-charts';
 import { Title } from '@patternfly/react-core';
 import { default as ChartTheme } from 'components/charts/chartTheme';
@@ -14,12 +15,13 @@ import { getCostRangeString } from 'components/charts/common/chartDatumUtils';
 import { getDateRange } from 'components/charts/common/chartDatumUtils';
 import {
   ChartSeries,
+  getChartNames,
   getDomain,
-  getEvents,
   getLegendData,
   getTooltipLabel,
   initHiddenSeries,
   isDataAvailable,
+  isSeriesHidden,
 } from 'components/charts/common/chartUtils';
 import getDate from 'date-fns/get_date';
 import i18next from 'i18next';
@@ -231,6 +233,19 @@ class HistoricalCostChart extends React.Component<HistoricalCostChartProps, Stat
     return currentRequestDate > 0 || previousRequestDate > 0 ? Math.max(currentRequestDate, previousRequestDate) : 31;
   }
 
+  // Returns onMouseOver, onMouseOut, and onClick events for the interactive legend
+  private getEvents() {
+    const { hiddenSeries, series } = this.state;
+
+    const result = getInteractiveLegendEvents({
+      chartNames: getChartNames(series),
+      isHidden: index => isSeriesHidden(hiddenSeries, index),
+      legendName: 'legend',
+      onLegendClick: props => this.handleLegendClick(props.index),
+    });
+    return result;
+  }
+
   private getLegend = () => {
     const { legendItemsPerRow } = this.props;
     const { hiddenSeries, series, width } = this.state;
@@ -247,8 +262,6 @@ class HistoricalCostChart extends React.Component<HistoricalCostChartProps, Stat
       />
     );
   };
-
-  // Interactive legend
 
   // Hide each data series individually
   private handleLegendClick = (index: number) => {
@@ -306,7 +319,7 @@ class HistoricalCostChart extends React.Component<HistoricalCostChartProps, Stat
             <Chart
               containerComponent={container}
               domain={domain}
-              events={getEvents(series, hiddenSeries, this.handleLegendClick)}
+              events={this.getEvents()}
               height={height}
               legendComponent={this.getLegend()}
               legendData={getLegendData(series, hiddenSeries)}
