+++ src/components/charts/historicalTrendChart/historicalTrendChart.tsx
@@ -7,18 +7,20 @@ import {
   ChartLegend,
   ChartLegendTooltip,
   createContainer,
+  getInteractiveLegendEvents,
 } from '@patternfly/react-charts';
 import { Title } from '@patternfly/react-core';
 import { default as ChartTheme } from 'components/charts/chartTheme';
 import { getCostRangeString, getDateRange } from 'components/charts/common/chartDatumUtils';
 import {
   ChartSeries,
+  getChartNames,
   getDomain,
-  getEvents,
   getLegendData,
   getTooltipLabel,
   initHiddenSeries,
   isDataAvailable,
+  isSeriesHidden,
 } from 'components/charts/common/chartUtils';
 import getDate from 'date-fns/get_date';
 import i18next from 'i18next';
@@ -177,6 +179,19 @@ class HistoricalTrendChart extends React.Component<HistoricalTrendChartProps, St
     return currentDate > 0 || previousDate > 0 ? Math.max(currentDate, previousDate) : 31;
   }
 
+  // Returns onMouseOver, onMouseOut, and onClick events for the interactive legend
+  private getEvents() {
+    const { hiddenSeries, series } = this.state;
+
+    const result = getInteractiveLegendEvents({
+      chartNames: getChartNames(series),
+      isHidden: index => isSeriesHidden(hiddenSeries, index),
+      legendName: 'legend',
+      onLegendClick: props => this.handleLegendClick(props.index),
+    });
+    return result;
+  }
+
   private getLegend = () => {
     const { legendItemsPerRow } = this.props;
     const { hiddenSeries, series } = this.state;
@@ -241,7 +256,7 @@ class HistoricalTrendChart extends React.Component<HistoricalTrendChartProps, St
             <Chart
               containerComponent={container}
               domain={domain}
-              events={getEvents(series, hiddenSeries, this.handleLegendClick)}
+              events={this.getEvents()}
               height={height}
               legendComponent={this.getLegend()}
               legendData={getLegendData(series, hiddenSeries)}
